requirements.txt
aws-cdk-lib==2.64.0
constructs>=10.0.0,<11.0.0
line-bot-sdk >=2.4.1
websocket-client >= 1.5.1
boto3 >=1.26.69
aws-solutions-constructs.aws-cloudfront-apigateway-lambda==2.30.0
aws-cdk.aws-lambda-python-alpha ==2.64.0-alpha.0